requirements.txt
biopython==1.79
certifi==2021.10.8
click==8.0.4
numpy==1.22.3
pandas==1.4.1
python-dateutil==2.8.2
pytz==2022.1
six==1.16.0
torch==1.10.0
typing_extensions==4.1.1
